Refactoring IBaseService implementation names - moving each implementation to separate directory with same file name BaseService
