+++ ConfluenceRESTClient/ConfluenceRESTClient.csproj
@@ -77,8 +77,8 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="Service\Base\Implementation\RestSharpBaseService.cs" />
-    <Compile Include="Service\Base\Implementation\DevDefinedBaseService.cs" />
+    <Compile Include="Service\Base\Implementation\RestSharp\BaseService.cs" />
+    <Compile Include="Service\Base\Implementation\DevDefined\BaseService.cs" />
     <Compile Include="Service\Base\IBaseService.cs" />
     <Compile Include="Service\Implementation\OAuthService.cs" />
     <Compile Include="Service\IOAuthService.cs" />
